move tmpl function into a package
